requirements.txt
dashscope==1.16.0
google-generativeai==0.4.1
openai==1.14.3
Pillow==10.2.0
python-telegram-bot==21.0.1
rich==13.7.1
